requirements/base.txt
Django==1.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-auth-lti==0.8
django-braces==1.3.1
django-canvas-api-token==0.1.1
django-dotenv==1.3.0
django-getenv==1.3.1
gunicorn==19.0.0
requests==2.5.0
git+https://github.com/Harvard-University-iCommons/django-auth-lti.git
